+++ app/Http/Controllers/AreasController.php
@@ -18,14 +18,20 @@ class AreasController extends Controller
      */
     public function index(Request $request)
     {
-        $areas = Area::paginate(20);
-        $data = $request->all();
+        $user = Auth::user();
+        if($user->admin == 1){
+            $areas = Area::paginate(20);
+            $data = $request->all();
 
-        if(isset($data['area_search'])){
-            $areas = Area::where('area_en','like',$data['area_search'])
-                ->orWhere('area_ar','like',$data['area_search'])->paginate(20);
+            if(isset($data['area_search'])){
+                $areas = Area::where('area_en','like',$data['area_search'])
+                    ->orWhere('area_ar','like',$data['area_search'])->paginate(20);
+            }
+            return view('areas', ['areas' => $areas]);
+        }else{
+            return redirect('/');
         }
-        return view('areas', ['areas' => $areas]);
+
 
     }
 
@@ -47,11 +53,18 @@ class AreasController extends Controller
      */
     public function store(AreaRequest $request)
     {
-        $areas = new Area();
-        $areas->area_en = $request->input('area_en');
-        $areas->area_ar = $request->input('area_ar');
-        $areas->save();
-        return redirect('/areas');
+        $user = Auth::user();
+        if($user->admin == 1){
+            $areas = new Area();
+            $areas->area_en = $request->input('area_en');
+            $areas->area_ar = $request->input('area_ar');
+            $areas->save();
+            return redirect('/areas');
+        }else{
+            return redirect('/');
+        }
+
+
     }
 
     /**
@@ -74,11 +87,16 @@ class AreasController extends Controller
      */
     public function update(AreaRequest $request, $id)
     {
-        $area = Area::find($id);
-        $area->area_en = $request->input('area_en');
-        $area->area_ar = $request->input('area_ar');
-        $area->save();
-        return redirect('/areas');
+        $user = Auth::user();
+        if($user->admin == 1){
+            $area = Area::find($id);
+            $area->area_en = $request->input('area_en');
+            $area->area_ar = $request->input('area_ar');
+            $area->save();
+            return redirect('/areas');
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -100,19 +118,22 @@ class AreasController extends Controller
      */
     public function deleteAreas(Request $request)
     {
-        $id = $request->get('id');
-        $areas = Area::with('restaurant')->where('id',$id)->get();
-        $images = [];
-        foreach ($areas as $area) {
-            foreach($area->restaurant as $restaurant){
-                $images[] = public_path('images/' . $restaurant->restaurant_image);
+        $user = Auth::user();
+        if($user->admin == 1){
+            $id = $request->get('id');
+            $areas = Area::with('restaurant')->where('id',$id)->get();
+            $images = [];
+            foreach ($areas as $area) {
+                foreach($area->restaurant as $restaurant){
+                    $images[] = public_path('images/' . $restaurant->restaurant_image);
+                }
             }
+            File::delete($images);
+            Area::whereIn('id',$id)->delete();
+            return redirect('/areas');
+        }else{
+            return redirect('/');
         }
-        File::delete($images);
-        Area::whereIn('id',$id)->delete();
-
-
-        return redirect('/areas');
     }
